requirements.txt
Django~=1.11
django-environ~=0.4.3
social-auth-app-django~=1.1.0
psycopg2~=2.7.3.1
